+++ CHANGELOG.md
@@ -16,11 +16,22 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2023-08-10
+
+- @triton-one/yellowstone-grpc:0.2.1
+- yellowstone-grpc-client-1.9.0+solana.1.16.1
+- yellowstone-grpc-geyser-1.7.0+solana.1.16.1
+- yellowstone-grpc-proto-1.9.0+solana.1.16.1
+
+### Features
+
+geyser: include entries to block message ([#169](https://github.com/rpcpool/yellowstone-grpc/pull/169)).
+
 ## 2023-07-26
 
 - yellowstone-grpc-geyser-1.6.1+solana.1.16.1
 
-### Fix
+### Fixes
 
 geyser: fix config example ([#168](https://github.com/rpcpool/yellowstone-grpc/pull/168)).